Move fee policy database structs
