+++ crates/database/src/lib.rs
@@ -5,6 +5,7 @@ pub mod auction_prices;
 pub mod byte_array;
 pub mod ethflow_orders;
 pub mod events;
+pub mod fee_policies;
 pub mod onchain_broadcasted_orders;
 pub mod onchain_invalidations;
 pub mod order_events;
